+++ src/features/linked-space/list/components/AllSpaceItem.tsx
